Display only running VMs in virt overview chart
