+++ frontend/packages/kubevirt-plugin/src/components/virtualization-overview/running-vms-per-template-card/RunningVMsChartLegendLabel.tsx
@@ -15,15 +15,16 @@ type RunningVMsChartLegendLabelProps = {
 };
 
 export const RunningVMsChartLegendLabel: React.FC<RunningVMsChartLegendLabelProps> = ({ item }) => {
-  const iconStyle = { color: item.color };
-  // const linkPath = `/k8s/ns/${item.namespace}/kubevirt.io~v1~VirtualMachine?labels=vm.kubevirt.io/template=${item.name}`
-  const linkPath = `/k8s/ns/${item.namespace}/virtualmachinetemplates/${item.name}`;
+  const iconStyle = { color: item?.color };
+  const linkPath = item?.namespace
+    ? `/k8s/ns/${item.namespace}/virtualmachinetemplates/${item?.name}`
+    : null;
 
   return (
     <>
       <i className="fas fa-square kv-running-vms-card__legend-label--color" style={iconStyle} />
-      <span className="kv-running-vms-card__legend-label--count">{item.vmCount}</span>{' '}
-      <Link to={linkPath}>{item.name}</Link>
+      <span className="kv-running-vms-card__legend-label--count">{item?.vmCount}</span>{' '}
+      {linkPath ? <Link to={linkPath}>{item?.name}</Link> : <span>{item?.name}</span>}
     </>
   );
 };
