+++ frontend/packages/kubevirt-plugin/src/components/virtualization-overview/running-vms-per-template-card/RunningVMsPerTemplateCard.tsx
@@ -2,89 +2,15 @@ import * as React from 'react';
 import { ChartDonut } from '@patternfly/react-charts';
 import { Card, CardBody, CardHeader, CardTitle, TitleSizes } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
-import { LABEL_USED_TEMPLATE_NAME } from '../../../constants';
-import { useRunningVMsPerTemplateResources } from '../../../hooks/use-running-vms-per-template-resources';
-import { getName, getNamespace } from '../../../selectors';
+import { useRunningVMsPerTemplateChartData } from '../../../hooks/useRunningVMsPerTemplateChartData';
 import { EmptyStateNoVMs } from '../../EmptyState/EmptyStateNoVMs';
 import { VMsChartLegend } from './RunningVMsChartLegend';
-import { RunningVMsChartLegendLabelItem } from './RunningVMsChartLegendLabel';
-import { getColorList } from './utils';
 
 import './running-vms-per-template-card.scss';
 
-const getTemplateNS = (templateName, templates) => {
-  const template = templates.find((temp) => getName(temp) === templateName);
-  return template ? getNamespace(template) : null;
-};
-
-const getTemplateToVMCountMap = (resources) => {
-  const loaded = resources?.loaded;
-  const vms = loaded && resources?.vms;
-  const templates = loaded && resources?.templates;
-
-  const templateToVMCountMap = new Map();
-  const numVMs = vms?.length || 0;
-
-  if (loaded) {
-    vms.forEach((vm) => {
-      const labels = vm?.metadata?.labels;
-      const template = labels ? labels[LABEL_USED_TEMPLATE_NAME] : 'Other';
-      const value = templateToVMCountMap.has(template)
-        ? templateToVMCountMap.get(template).vmCount + 1
-        : 1;
-      templateToVMCountMap.set(template, { vmCount: value });
-    });
-  }
-
-  const numTemplates = templateToVMCountMap.size;
-  const colorListIter = getColorList(numTemplates).values();
-
-  for (const key of templateToVMCountMap.keys()) {
-    const templateChartData = templateToVMCountMap.get(key);
-    const additionalData = {
-      percentage: numVMs ? Math.round((templateChartData.vmCount / numVMs) * 100) : 0,
-      color: colorListIter.next().value,
-      namespace: getTemplateNS(key, templates),
-    };
-    templateToVMCountMap.set(key, { ...templateChartData, ...additionalData });
-  }
-
-  return templateToVMCountMap;
-};
-
-const getChartData = (templateToVMCountMap) => {
-  const chartData = [];
-  templateToVMCountMap.forEach((data, templateName) => {
-    chartData.push({
-      x: templateName,
-      y: data.percentage,
-      fill: data.color,
-    });
-  });
-  return chartData;
-};
-
-const getLegendItems = (templateToVMCountMap): RunningVMsChartLegendLabelItem[] => {
-  const legendItems = [];
-  templateToVMCountMap.forEach((data, templateName) => {
-    legendItems.push({
-      name: templateName,
-      vmCount: data.vmCount,
-      color: data.color,
-      namespace: data.namespace,
-    });
-  });
-  return legendItems;
-};
-
 export const RunningVMsPerTemplateCard = () => {
   const { t } = useTranslation();
-  const resources = useRunningVMsPerTemplateResources();
-  const templateToVMCountMap = React.useMemo(() => getTemplateToVMCountMap(resources), [resources]);
-
-  const chartData = getChartData(templateToVMCountMap);
-  const legendItems = getLegendItems(templateToVMCountMap);
-  const numVMs = resources?.vms?.length;
+  const [chartData, legendItems, numVMs] = useRunningVMsPerTemplateChartData();
 
   const chart = (
     <div>
